WIP - need to figure out how to better control the python instance running my script

+++ Source/GUI/FolderOrganizer/BackEnd/ScriptWrapper.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using FolderOrganizer.Application;
 using FolderOrganizer.Logging;
@@ -11,25 +12,45 @@ using FolderOrganizer.UILib;
 
 namespace FolderOrganizer.BackEnd
 {
-    internal struct RuntimeKeys
-    {
-        internal struct RuntimeInfo
-        {
-            public const string ProcessID = "pid";
-        }
-
-        internal struct Paths
-        {
-            public const string SourcePath = "Source";
-            public const string DestinationPath = "Destination";
-        }
-    }
 
     class ScriptWrapper
     {
         private static ScriptWrapper _instance = null;
 
         public static ScriptWrapper Instance => _instance ?? (_instance = new ScriptWrapper());
+
+        internal class RuntimePaths
+        {
+            private string _src;
+            private string _dest;
+
+            public string Source
+            {
+                get => _src;
+                set
+                {
+                    Logger.Inf($"Source path: \"{value}\"");
+                    _src = value;
+                }
+            }
+            public string Destination
+            {
+                get => _dest;
+                set
+                {
+                    Logger.Inf($"Destination path: \"{value}\"");
+                    _dest = value;
+                }
+            }
+
+            public Dictionary<string, string> AsDictionary()
+            {
+                var dict = new Dictionary<string, string>();
+                dict[ConfigurationKeys.Paths.SourcePath] = Source;
+                dict[ConfigurationKeys.Paths.DestinationPath] = Destination;
+                return dict;
+            }
+        }
 
         private enum RuntimeFlags
         {
@@ -39,6 +60,7 @@ namespace FolderOrganizer.BackEnd
 
         private Process _process;
         public readonly Dictionary<string, string> _runtimeInfo = new Dictionary<string, string>();
+        public RuntimePaths Paths { get; private set; }
 
         private readonly string _pathsFilePath;
         private readonly string _scriptFilePath;
@@ -49,10 +71,17 @@ namespace FolderOrganizer.BackEnd
             _pathsFilePath = Path.Combine(AppFolders.ConfigsDir, "Paths.ini");
             _scriptFilePath = Path.Combine(AppFolders.ScriptsDir, "FolderOrganizer.py");
             _runtimeInfoFilePath = Path.Combine(AppFolders.ConfigsDir, "RuntimeInfo.ini");
+            Paths = new RuntimePaths();
 
             LoadFromDisk();
 
             Logger.Inf($"Script filename: {_scriptFilePath}");
+        }
+
+        public void SetRuntimePaths(string src, string dest)
+        {
+            Paths.Source = src;
+            Paths.Destination = dest;
         }
 
         public void Launch()
@@ -68,8 +97,24 @@ namespace FolderOrganizer.BackEnd
 
         public void Terminate()
         {
-            RaiseFlag(RuntimeFlags.Terminate);
+            var flag = RuntimeFlags.Terminate;
+            RaiseFlag(flag);
+            var flagFilePath = Path.Combine(AppFolders.FlagsDir, flag.ToString());
+            var timeoutCounter = 10;
+            while (File.Exists(flagFilePath) && timeoutCounter > 0)
+            {
+                Thread.Sleep(1000);
+                timeoutCounter--;
+            }
+            ForceClose();
         }
+
+        void ForceClose()
+        {
+            Logger.Inf("Force closing python");
+            _process.Close();
+        }
+
         void RaiseFlag(RuntimeFlags flag)
         {
 
@@ -77,6 +122,7 @@ namespace FolderOrganizer.BackEnd
             using (var file = File.Create(flagFilePath))
             {
                 Logger.Inf($"Raising flag: {flag}");
+                Logger.Dbg($"Flag file path: {flagFilePath}");
             }
         }
 
@@ -87,6 +133,8 @@ namespace FolderOrganizer.BackEnd
 
         void LaunchImpl()
         {
+            IniFile.EditFile(_pathsFilePath, Paths.AsDictionary());
+
             Logger.Bnr("Launching script", "*", 5);
 
             var scriptFilePath = Path.Combine(AppFolders.ScriptsDir, _scriptFilePath);
@@ -96,14 +144,13 @@ namespace FolderOrganizer.BackEnd
             if (!File.Exists(scriptFilePath))
                 Logger.Ftl("Script doesn't exist!");
 
-            // var scriptEngine = IronPython.Hosting.Python.CreateEngine();
-            // scriptEngine.ExecuteFile(_scriptFilePath);
-            
+            var pythonExePath = PythonExe.GetVersionPath(PythonExe.HighestVersion);
+
             _process = new Process
             {
                 StartInfo =
                 {
-                    FileName = "C:\\Users\\44753\\AppData\\Local\\Programs\\Python\\Python39\\python.exe",
+                    FileName = pythonExePath,
                     Arguments = $"{_scriptFilePath} ",
                     UseShellExecute = false,
                     RedirectStandardError = true,
@@ -112,17 +159,22 @@ namespace FolderOrganizer.BackEnd
                 }
             };
             _process.Start();
+            
+            _runtimeInfo[ConfigurationKeys.RuntimeInfo.ProcessID] = _process.Id.ToString();
 
-            var pid = _process.Id;
-            _runtimeInfo[RuntimeKeys.RuntimeInfo.ProcessID] = pid.ToString();
+            var processes = Process.GetProcesses();
+
             Logger.Bnr("Script launched", "*", 5);
         }
 
         void StoreToDisk()
         {
             Logger.Bnr("Storing to disk", "*", 5);
-            IniFile.DeleteFile(_runtimeInfoFilePath);
-            IniFile.WriteFile(_runtimeInfoFilePath, _runtimeInfo);
+
+            if (File.Exists(_runtimeInfoFilePath))
+                IniFile.EditFile(_runtimeInfoFilePath, _runtimeInfo);
+            else
+                IniFile.WriteFile(_runtimeInfoFilePath, _runtimeInfo);
             Logger.Bnr("Store concluded", "*", 5);
         }
 
@@ -143,7 +195,7 @@ namespace FolderOrganizer.BackEnd
             if (!IniFile.ReadFile(_runtimeInfoFilePath, _runtimeInfo))
                 return;
 
-            if (!_runtimeInfo.TryGetValue(RuntimeKeys.RuntimeInfo.ProcessID, out var pidStr))
+            if (!_runtimeInfo.TryGetValue(ConfigurationKeys.RuntimeInfo.ProcessID, out var pidStr))
                 return;
 
             var pid = int.Parse(pidStr);
@@ -166,25 +218,30 @@ namespace FolderOrganizer.BackEnd
 
         void LoadPaths()
         {
-            if (!IniFile.ReadFile(_pathsFilePath, _runtimeInfo))
+            Dictionary<string, string> pathsIniData = new Dictionary<string, string>();
+            if (!IniFile.ReadFile(_pathsFilePath, pathsIniData))
                 return;
 
-            if (_runtimeInfo.TryGetValue(RuntimeKeys.Paths.SourcePath, out var sourcePath))
             {
-                Logger.Inf($"Source path found: \"{sourcePath}\"");
-            }
-            else
-            {
-                Logger.Wrn("Source path not found");
+                if (pathsIniData.TryGetValue(ConfigurationKeys.Paths.SourcePath, out var sourcePath))
+                {
+                    Paths.Source = sourcePath;
+                }
+                else
+                {
+                    Logger.Wrn("Source path not found");
+                }
             }
 
-            if (_runtimeInfo.TryGetValue(RuntimeKeys.Paths.DestinationPath, out var destPath))
             {
-                Logger.Inf($"Destination path found: \"{destPath}\"");
-            }
-            else
-            {
-                Logger.Wrn("Destination path not found");
+                if (pathsIniData.TryGetValue(ConfigurationKeys.Paths.DestinationPath, out var destPath))
+                {
+                    Paths.Destination = destPath;
+                }
+                else
+                {
+                    Logger.Wrn("Destination path not found");
+                }
             }
 
         }
